Move code, fix linter
